# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: X.orb API
  version: '0.5.0'
  description: |
    X.orb — The orchestration layer for AI agent trust.
    8-gate security pipeline, reputation scoring, economic bonding, and x402 micropayments.

    All responses use a standard envelope: `{ success: boolean, data: T, pagination?: {...}, error?: { code, message } }`

    Free tier: 500 actions/month, no wallet required.
    Standard: 501+ actions, 30 basis points (0.30%) per action in USDC.
    High Volume: 50,001+ actions, 15 basis points (0.15%).
  contact:
    name: X.orb Support
    email: contact@xorb.xyz
    url: https://www.xorb.xyz
  license:
    name: Proprietary
    url: https://www.xorb.xyz/terms
servers:
  - url: https://api.xorb.xyz
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key starting with xorb_sk_. Obtain via POST /v1/auth/keys.
  headers:
...
    X-Content-Hash:
      schema: { type: string }
      description: SHA-256 hash of response content (for tamper evidence)
  schemas:
    Envelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        data: {}
        pagination:
          type: object
          properties:
            next_cursor: { type: string, nullable: true }
            has_more: { type: boolean }
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
    Agent:
      type: object
      required: [agentId, name, role, sponsorAddress, status, reputation]
      properties:
        agentId: { type: string, example: agent_a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4 }
        name: { type: string, example: research-bot }
        role: { type: string, enum: [TRADER, RESEARCHER, MONITOR, CODER, GOVERNANCE_ASSISTANT, FILE_INDEXER, RISK_AUDITOR] }
        sponsorAddress: { type: string }
        sessionWalletAddress: { type: string }
        stakeBond: { type: string, description: USDC in 6 decimals }
        reputation: { type: integer, minimum: 0, maximum: 10000 }
        reputationTier: { type: string, enum: [UNTRUSTED, NOVICE, RELIABLE, TRUSTED, ELITE] }
        status: { type: string, enum: [active, paused, revoked, expired] }
        createdAt: { type: integer }
        expiresAt: { type: integer }
        totalActionsExecuted: { type: integer }
        slashEvents: { type: integer }
        description: { type: string }
...
    CreateAgentRequest:
      type: object
      required: [name, role, sponsor_address]
      properties:
        name: { type: string, minLength: 2, maxLength: 64 }
        role: { type: string, enum: [TRADER, RESEARCHER, MONITOR, CODER, GOVERNANCE_ASSISTANT, FILE_INDEXER, RISK_AUDITOR] }
        sponsor_address: { type: string, minLength: 10 }
        stake_bond: { type: string }
        expiry_days: { type: integer, minimum: 0, maximum: 365 }
        description: { type: string, maxLength: 256 }
...
    WebhookSubscription:
      type: object
      properties:
        id: { type: string }
        url: { type: string, format: uri }
        event_types: { type: array, items: { type: string } }
        secret: { type: string }
        active: { type: boolean }
    PaymentRecord:
      type: object
      properties:
        id: { type: string }
        action_id: { type: string }
        agent_id: { type: string }
        gross_usdc: { type: string }
        fee_usdc: { type: string }
        net_usdc: { type: string }
        fee_bps: { type: integer }
        status: { type: string, enum: [held, completed, escrowed, refunded, failed] }
        collect_tx: { type: string, nullable: true }
        fee_tx: { type: string, nullable: true }
        forward_tx: { type: string, nullable: true }
        refund_tx: { type: string, nullable: true }
        chain: { type: string }
        created_at: { type: string, format: date-time }
    Error400:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - properties:
            success: { enum: [false] }
            error: { properties: { code: { example: validation_error }, message: { example: Name must be 2-64 characters } } }
    Error401:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - properties:
            error: { properties: { code: { example: invalid_api_key }, message: { example: Invalid API key } } }
    Error403:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - properties:
            error: { properties: { code: { example: forbidden }, message: { example: You do not own this agent } } }
    Error404:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - properties:
            error: { properties: { code: { example: not_found }, message: { example: Agent not found } } }
    Error429:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - properties:
            error: { properties: { code: { example: rate_limit_exceeded }, message: { example: Rate limit exceeded. Retry after 1 second. } } }
...
  responses:
    BadRequest:
      description: Validation error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error400' } } }
    Unauthorized:
      description: Missing or invalid API key
      content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } }
    Forbidden:
      description: Insufficient permissions
      content: { application/json: { schema: { $ref: '#/components/schemas/Error403' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error404' } } }
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error429' } } }
...
  parameters:
    x-payment:
      name: x-payment
      in: header
      description: Base64-encoded x402 payment header
      schema: { type: string }
...
paths:
...
  /v1/agents:
    post:
      summary: Register a new agent
      tags: [Agents]
      parameters: [{ $ref: '#/components/parameters/x-payment' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAgentRequest' }
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { type: object, properties: { agent: { $ref: '#/components/schemas/Agent' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { description: Free tier exhausted }
        '429': { $ref: '#/components/responses/RateLimited' }
    get:
      summary: List agents (owned by authenticated sponsor)
      tags: [Agents]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
      responses:
        '200':
          description: Agent list with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/Agent' } }
                  pagination: { type: object, properties: { has_more: { type: boolean } } }
                  total: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/agents/{id}:
    get:
      summary: Get agent details
      tags: [Agents]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Agent details }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update agent (pause/resume/renew)
      tags: [Agents]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [pause, resume, renew] }
                renew_days: { type: integer, minimum: 1, maximum: 365 }
      responses:
        '200': { description: Agent updated }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Revoke agent permanently
      tags: [Agents]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Agent revoked }
  /v1/agents/{id}/public:
    get:
      summary: Public agent profile (no auth)
      security: []
      tags: [Agents]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Public profile (name, role, reputation, status) }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/audit/{agentId}:
    get:
      summary: Get audit log
      tags: [Audit]
      parameters: [{ name: agentId, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Audit log with events, violations, reputation }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/audit/{actionId}/evidence:
    post:
      summary: Upload audit evidence (PDF/PNG/JPG, max 10MB)
      tags: [Audit]
      parameters: [{ name: actionId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '201': { description: Evidence uploaded }
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/audit/export/{agentId}:
    get:
      summary: Export audit log (CSV or PDF)
      tags: [Audit]
      parameters:
        - { name: agentId, in: path, required: true, schema: { type: string } }
        - { name: format, in: query, schema: { type: string, enum: [csv, pdf], default: csv } }
      responses:
        '200':
          description: File download
          headers:
            X-Content-Hash: { $ref: '#/components/headers/X-Content-Hash' }
...
  /v1/webhooks:
    post:
      summary: Subscribe to webhook events
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, event_types]
              properties:
                url: { type: string, format: uri }
                event_types: { type: array, items: { type: string }, minItems: 1, description: 'Event types: agent.registered, action.approved, action.blocked, agent.slashed, reputation.changed, payment.completed, payment.refunded, etc.' }
      responses:
        '201':
          description: Webhook created (includes signing secret)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/WebhookSubscription' }
    get:
      summary: List webhook subscriptions
      tags: [Webhooks]
      responses:
        '200': { description: Webhook list }
...
  /v1/events:
    get:
      summary: Get recent events (ownership-filtered)
      tags: [Events]
      parameters:
        - { name: agent_id, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
        - { name: since, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Events with cursor pagination
...
  /v1/payments:
    get:
      summary: Payment history for authenticated sponsor
      tags: [Payments]
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Payment list with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/PaymentRecord' } }
                  pagination: { type: object, properties: { next_cursor: { type: string }, has_more: { type: boolean } } }
...
  /v1/export/audit/{agentId}:
    get:
      summary: Export audit log as CSV
      tags: [Export]
      parameters: [{ name: agentId, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: CSV file download }
...